requirements.txt
Django==3.0.7
psycopg2
flake8>=3.7.0,<3.8.0
djangorestframework
django-rest-auth